controllers: refactor imports for repos
